backend/requirements.txt
Flask
pymongo
requests
pdfplumber
python-docx
langdetect
googletrans==4.0.0-rc1
Flask-Cors
python-dotenv
sumy
nltk